+++ src/Objetos/Turno.cs
@@ -13,14 +13,14 @@ namespace UberFrba.Objetos
         public int hora_inicio { get; set; }
         public int hora_fin { get; set; }
         public string descripcion { get; set; }
-        public int valor_km { get; set; }
-        public int precio_base { get; set; }
+        public double valor_km { get; set; }
+        public double precio_base { get; set; }
         public bool habilitado { get; set; }
         
         public Turno() { }
 
         //Este turno se crea cuando se hace una alta de turno
-        public Turno (int horaInicio, int horaFin, string desc, int valorKm, int precioBase, bool hab) {
+        public Turno (int horaInicio, int horaFin, string desc, double valorKm, double precioBase, bool hab) {
             this.hora_inicio = horaInicio;
             this.hora_fin = horaFin;
             this.descripcion = desc;
@@ -29,7 +29,7 @@ namespace UberFrba.Objetos
             this.habilitado = hab;
         }
 
-        public Turno(int id,int horaInicio, int horaFin, string desc, int valorKm, int precioBase, bool hab)
+        public Turno(int id,int horaInicio, int horaFin, string desc, double valorKm, double precioBase, bool hab)
         {
             this.id_turno = id;
             this.hora_inicio = horaInicio;
@@ -46,8 +46,8 @@ namespace UberFrba.Objetos
             this.hora_inicio = Int32.Parse(datosTurno.Cells["hora_inicio"].Value.ToString());
             this.hora_fin = Int32.Parse(datosTurno.Cells["hora_fin"].Value.ToString());
             this.descripcion = datosTurno.Cells["descripcion"].Value.ToString();
-            this.valor_km = Int32.Parse(datosTurno.Cells["valor_km"].Value.ToString());
-            this.precio_base = Int32.Parse(datosTurno.Cells["precio_base"].Value.ToString());
+            this.valor_km = Double.Parse(datosTurno.Cells["valor_km"].Value.ToString());
+            this.precio_base = Double.Parse(datosTurno.Cells["precio_base"].Value.ToString());
             this.habilitado = Convert.ToBoolean(datosTurno.Cells["habilitado"].Value.ToString());
         }
